# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces/{ws}/relations": {
      "post": {
        "operationId": "RelationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRelationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a relation — needs creator on BOTH databases",
        "tags": [
          "relations"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/relations/{rel}": {
      "delete": {
        "operationId": "RelationsController_remove",
        "parameters": [
          {
            "name": "rel",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteRelationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a relation, both its fields, and all links (confirm: true)",
        "tags": [
          "relations"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/records/{rec}/links/{field}": {
      "get": {
        "operationId": "LinksController_list",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Linked records for a relation field ({id, title} chips)",
        "tags": [
          "relations"
        ]
      },
      "post": {
        "operationId": "LinksController_add",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkRecordsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add links (409 when one-to-many already linked)",
        "tags": [
          "relations"
        ]
      },
      "put": {
        "operationId": "LinksController_replace",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceLinksDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace all links for this record on this field",
        "tags": [
          "relations"
        ]
      },
      "delete": {
        "operationId": "LinksController_remove",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rec",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkRecordsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove specific links",
        "tags": [
          "relations"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/views": {
      "post": {
        "operationId": "ViewsController_create",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateViewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a saved view (config validated against live fields)",
        "tags": [
          "views"
        ]
      }
    },
    "/api/v1/workspaces/{ws}/databases/{db}/views/{view}": {
      "patch": {
        "operationId": "ViewsController_update",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "view",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateViewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rename / reconfigure / reorder a view",
        "tags": [
          "views"
        ]
      },
      "delete": {
        "operationId": "ViewsController_remove",
        "parameters": [
          {
            "name": "db",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "view",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a view (409 on the last one)",
        "tags": [
          "views"
        ]
      }
    },
...
  },
  "info": {
    "title": "StoryOS API",
    "description": "The public StoryOS REST API. The web app is client #1 of this same API — everything the UI does, you can do.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateRelationDto": {
        "type": "object",
        "properties": {
          "database_a_id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "database_b_id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "cardinality": {
            "type": "string",
            "enum": [
              "one_to_many",
              "many_to_many"
            ]
          },
          "field_a_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "field_b_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "database_a_id",
          "database_b_id",
          "cardinality"
        ]
      },
      "DeleteRelationDto": {
        "type": "object",
        "properties": {
          "confirm": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "confirm"
        ]
      },
      "LinkRecordsDto": {
        "type": "object",
        "properties": {
          "record_ids": {
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        },
        "required": [
          "record_ids"
        ]
      },
      "ReplaceLinksDto": {
        "type": "object",
        "properties": {
          "record_ids": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        },
        "required": [
          "record_ids"
        ]
      },
      "CreateViewDto__schema0": {
        "anyOf": [
          {
            "type": "object",
            "properties": {
              "field": {
                "type": "string",
                "minLength": 1
              },
              "op": {
                "type": "string",
                "enum": [
                  "eq",
                  "neq",
                  "contains",
                  "gt",
                  "gte",
                  "lt",
                  "lte",
                  "before",
                  "after",
                  "within",
                  "has",
                  "has_none",
                  "is_empty",
                  "not_empty"
                ]
              },
              "value": {}
            },
            "required": [
              "field",
              "op"
            ]
          },
          {
            "type": "object",
            "properties": {
              "and": {
                "minItems": 1,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateViewDto__schema0"
                }
              }
            },
            "required": [
              "and"
            ]
          },
          {
            "type": "object",
            "properties": {
              "or": {
                "minItems": 1,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateViewDto__schema0"
                }
              }
            },
            "required": [
              "or"
            ]
          }
        ]
      },
      "CreateViewDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "enum": [
              "table",
              "board"
            ]
          },
          "config": {
            "default": {
              "sorts": [],
              "hidden_field_ids": [],
              "card_field_ids": [],
              "column_widths": {}
            },
            "type": "object",
            "properties": {
              "filters": {
                "$ref": "#/components/schemas/CreateViewDto__schema0"
              },
              "sorts": {
                "default": [],
                "maxItems": 3,
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string",
                      "minLength": 1
                    },
                    "direction": {
                      "default": "asc",
                      "type": "string",
                      "enum": [
                        "asc",
                        "desc"
                      ]
                    }
                  },
                  "required": [
                    "field"
                  ]
                }
              },
              "hidden_field_ids": {
                "default": [],
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                }
              },
              "group_by_field_id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "card_field_ids": {
                "default": [],
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                }
              },
              "column_widths": {
                "default": {},
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "minimum": 40,
                  "maximum": 1200
                }
              }
            }
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "UpdateViewDto__schema0": {
        "anyOf": [
          {
            "type": "object",
            "properties": {
              "field": {
                "type": "string",
                "minLength": 1
              },
              "op": {
                "type": "string",
                "enum": [
                  "eq",
                  "neq",
                  "contains",
                  "gt",
                  "gte",
                  "lt",
                  "lte",
                  "before",
                  "after",
                  "within",
                  "has",
                  "has_none",
                  "is_empty",
                  "not_empty"
                ]
              },
              "value": {}
            },
            "required": [
              "field",
              "op"
            ]
          },
          {
            "type": "object",
            "properties": {
              "and": {
                "minItems": 1,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UpdateViewDto__schema0"
                }
              }
            },
            "required": [
              "and"
            ]
          },
          {
            "type": "object",
            "properties": {
              "or": {
                "minItems": 1,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UpdateViewDto__schema0"
                }
              }
            },
            "required": [
              "or"
            ]
          }
        ]
      },
      "UpdateViewDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "config": {
            "type": "object",
            "properties": {
              "filters": {
                "$ref": "#/components/schemas/UpdateViewDto__schema0"
              },
              "sorts": {
                "default": [],
                "maxItems": 3,
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string",
                      "minLength": 1
                    },
                    "direction": {
                      "default": "asc",
                      "type": "string",
                      "enum": [
                        "asc",
                        "desc"
                      ]
                    }
                  },
                  "required": [
                    "field"
                  ]
                }
              },
              "hidden_field_ids": {
                "default": [],
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                }
              },
              "group_by_field_id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "card_field_ids": {
                "default": [],
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                }
              },
              "column_widths": {
                "default": {},
                "type": "object",
                "additionalProperties": {
                  "type": "integer",
                  "minimum": 40,
                  "maximum": 1200
                }
              }
            }
          },
          "position": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        }
      },
...
    }
  }
}